pyproject.toml
[tool.poetry]
name = "ohdsi-webapi-client"
version = "0.1.0"
description = "Python client for OHDSI WebAPI (MVP: info, sources, vocabulary, concept sets, cohorts)."
authors = ["Your Name <you@example.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{ include = "ohdsi_webapi", from = "src" }]
keywords = ["OHDSI", "WebAPI", "OMOP", "ETL", "Cohort", "Healthcare"]
homepage = "https://github.com/your-org/ohdsi-webapi-client"
repository = "https://github.com/your-org/ohdsi-webapi-client"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: Apache Software License",
  "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
httpx = { version = ">=0.27,<0.28", extras=["http2"] }
pydantic = ">=2.7,<3.0"
tenacity = ">=8.2,<9.0"
typing-extensions = ">=4.11,<5.0"
python-dotenv = "^1.1.1"

[tool.poetry.extras]
cli = ["typer", "rich"]
pandas = ["pandas"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
respx = "^0.20.2"
ruff = "^0.4.0"
mypy = "^1.10.0"
black = "^24.4.0"
pre-commit = "^3.7.0"
vcrpy = "^6.0.0"
ipython = ">=8.0,<9.0"

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B", "UP", "PT", "RUF"]

[tool.black]
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
show_error_codes = true

[tool.pytest.ini_options]
addopts = "-q --cov=ohdsi_webapi --cov-report=term-missing"

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"
